Create FlashCard.py
A flashcard is a card having information on both sides, which can be used as an aid in memoization. Flashcards usually have a question on one side and an answer on the other. Particularly in this file, we are going to create flashcards that will be having a word and its meaning.